+++ src/main/java/subway/line/presentation/LineController.java
@@ -1,8 +1,9 @@
-package subway.line;
+package subway.line.presentation;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import subway.line.service.LineService;
 
 import java.net.URI;
 import java.util.ArrayList;
