+++ src/main/java/subway/station/domain/Station.java
@@ -1,10 +1,11 @@
-package subway.station;
+package subway.station.domain;
 
 public class Station {
     private Long id;
     private String name;
 
-    public Station(){}
+    public Station() {
+    }
 
     public Station(Long id, String name) {
         this.id = id;
