+++ src/test/java/subway/line/SectionAcceptanceTest.java
@@ -9,8 +9,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import subway.AcceptanceTest;
-import subway.section.SectionRequest;
-import subway.station.StationResponse;
+import subway.section.presentation.SectionRequest;
+import subway.station.presentation.StationResponse;
+import subway.line.presentation.LineResponse;
 
 import java.util.Arrays;
 import java.util.List;
@@ -124,7 +125,7 @@ public class SectionAcceptanceTest extends AcceptanceTest {
         지하철_구간_생성_요청(lineResponse, upStation, downStation, distance);
     }
 
-    public static ExtractableResponse<Response> 지하철_구간_생성_요청(LineResponse line,
+    public static ExtractableResponse<Response> 지하철_구간_생성_요청(subway.line.presentation.LineResponse line,
                                                              StationResponse upStation,
                                                              StationResponse downStation,
                                                              int distance) {
